do listcommand command

+++ chat/client/src/main/java/client/view/chat_session/events/FileMetadata.java
@@ -1,19 +1,12 @@
 package client.view.chat_session.events;
 
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.layout.HBox;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
 public class FileMetadata {
   private final String fileId;
   private final String fileName;
-  private final int size;
+  private final long size;
   private final String mimeType;
 
-  public FileMetadata(String fileId, String fileName, int size, String mimeType) {
+  public FileMetadata(String fileId, String fileName, long size, String mimeType) {
     this.fileId = fileId;
     this.fileName = fileName;
     this.mimeType = mimeType;
@@ -28,7 +21,7 @@ public class FileMetadata {
     return fileName;
   }
 
-  public int getSize() {
+  public long getSize() {
     return size;
   }
